Reduce dependencies on external binaries.
This includes changes which greatly simplify the build system by
getting rid of the conditional support on Bash, when
Bash is needed for a succesful build anyway. Other changes
include using the "jar" binary (part of JDK) to manipulate jar's,
instead of the external unzip binary (not part of msysgit on
Windows). Issue 4052.

These changes need to be thorougly reviewed - they've not been
tested on Linux at all, and there are still issues with the build
on Windows.

Change-Id: Ib409a977462398156b92b49f824e773694e4f210