Reduce dependencies on external binaries.
This includes changes which greatly simplify the build system by
getting rid of the conditional support on Bash, when
Bash is needed for a succesful build anyway. Other changes
include using the "jar" binary (part of JDK) to manipulate jar's,
instead of the external unzip binary (not part of msysgit on
Windows). Issue 4052.

Change-Id: Ib409a977462398156b92b49f824e773694e4f210
Reviewed-on: http://gerrit.dmdirc.com/1164
Automatic-Compile: DMDirc Local Commits <dmdirc@googlemail.com>
Reviewed-by: Shane Mc Cormack <shane@dmdirc.com>