Fix for issue CLIENT-312 - Parser shouldn't extend runnable
The parser now defines some new methods:
  - connect()
  - connect(threadName)
  - interrupt()

connect([threadName]) is responsible for creating any required threads
for the parser to run in and maintaining them internally.
interrupt() allows for interrupting these thread if they exist and if
required.

A new base class for parsers has been added to accomodate this with
minimum changes required elsewhere - ThreadedParser which BaseParser
now extends.

ThreadedParser handles the creation of a single controlThread and allows
interrupting it, and also defines an abstract void run() method (which
all existing parsers will already have from extending runnable in the
past) which is used by the controlThread.

Change-Id: I5ef8ea1e988fe9591df0054ad0016c409fc34b3f
Depends-On: Ib3e884407181adc362394b23fe59f8e3b401d692
Depends-On: I90ac61d56eb623165d3c4211f1c3247fcba823c8
Reviewed-on: http://gerrit.dmdirc.com/2256
Automatic-Compile: DMDirc Build Manager
Reviewed-by: Chris Smith <chris@dmdirc.com>