Minor changes to facilitate testing.
Remove some 'final' modifiers for CommandParser methods.

Modify the Logger class so it only calls into the ErrorManager if
it actually needs it. The ErrorManager constructor requires an
IdentityManager which causes any test that executes a Logger.assert
line to require most of the client is constructed, regardless of
what the test is actually doing.

Change-Id: I63879db5c13517cfa60941fce73a2ba01e63f1cd
Reviewed-on: http://gerrit.dmdirc.com/2681
Reviewed-by: Greg Holmes <greg@dmdirc.com>
Automatic-Compile: DMDirc Build Manager