Add concept of Displayable events.
These are events which are going to cause a line to be added to a
client window. They allow actions, plugins, or other listeners
to change the formatter line.

Eventually the UI model will just be able to listen for displayable
events on the local event bus to figure out what needs to be displayed.

Change-Id: Ia86c593df66f8bff13574714c45464a2634c02ba
Reviewed-on: http://gerrit.dmdirc.com/3390
Automatic-Compile: DMDirc Build Manager
Reviewed-by: Greg Holmes <greg@dmdirc.com>