Add a way to get a FrameContainer from Connection.
At the moment Server is both a FrameContainer and a Connection, and
in most places we have to cast up to Server to do what we need to.

This adds a method to Connection to retrieve the associated
FrameContainer. Eventually they should be separate objects, with
the Connection implementation managing the parser and associated
state, and the FrameContainer modelling the window (providing
tab completers, line history, input handlers, etc).

Change-Id: I56c9a2bb9ad06e9e3ad911914b29a046959e30ea
Reviewed-on: http://gerrit.dmdirc.com/3431
Automatic-Compile: DMDirc Build Manager
Reviewed-by: Greg Holmes <greg@dmdirc.com>