Add CoreProfilesDialogModel.
The profiles wrapper needs to be replaced (probably), the model is
flexible enough to cope with this. It's still a little bit hacky
with the wrapped model and if we stick with them we should probably
change this model.

EditSelectedProfileNameValidator and EditSelectedNicknameValidator
are both a work in progress and quite like prevent the editing of
profiles.

This still needs to be unit tested.

Change-Id: I9dd73f62ae3641b8d11da9881ad13a11a5b4453b
Reviewed-on: http://gerrit.dmdirc.com/3617
Automatic-Compile: DMDirc Build Manager
Reviewed-by: Chris Smith <chris@dmdirc.com>