Point IDEA at directories not individual jars.
This replaces the idea libraries with jar directories.

Also creates a 'bundled' lib directory in the main client, which
contains the jars that end up bundled with DMDirc. This allows
the plugins project to cleanly depend on these without depending
on any other dependencies the client has.

Change-Id: I929f402ebef5344dbc133b9a9828868e4f8c2dcd
Reviewed-on: http://gerrit.dmdirc.com/3735
Automatic-Compile: DMDirc Build Manager
Reviewed-by: Greg Holmes <greg@dmdirc.com>